Resolve "Get OpenMP compilation working in PION with cmake"
+ Added OpenMP instructions to cmake files
+ Added OpenMP parallel execution for calculating timestep, calculating dynamics and microphysics updates, and updating cells.
+ Removed MP as global variable, and added it to setup_fixed_grid class.
+ set microphysics to inherit interpolation class instead of having an instance as member data
+ merged main_MPI into main, icgen_MPI into icgen
+ Added test problem for r-hd in 2D

Remaining issues:
+ radiative transfer is not working because the mpi ranks are not set correctly anymore (fix in new branch)
+ radiative transfer is not multi-threaded (fix in new branch, if possible)
+ 1D simulations are not multi-threaded because the threading is based on columns of cells (think about whether it is worthwhile)

+++ source/grid/uniform_grid.cpp
@@ -488,7 +488,7 @@ int UniformGrid::allocate_grid_data()
     cout << "i=" << i << " of " << G_ncell_all;
     cout << ": ix=" << ix << ", nel=" << nel << "\n";
 #endif
-    carr[i].id = i;
+    carr[i].id = static_cast<long int>(i);
     ix         = CI.set_cell_pointers(&(carr[i]), data, ix);
     if (i < G_ncell_all - 1)
       carr[i].npt_all = &(carr[i + 1]);
